caching with uuid
The wasm_exec.js and main.wasm files are requested with a uuid query param to ensure the version the server expects is used.
Inlining the run_wasm.js file so it requests the main.wasm file with the uuid query parameter.  Wrapping in a function to run after the document is loaded to ensure wasm_exce.js is loaded first.