Renaming/altering Board HasSingleUsedGroup()
Renaming Board.HasSingleUsedGroup to CanBeFinished().  In addition to boards that only have one group, false is returned if a board has zero groups, in addition to one group.  This simplifies game tests by allowing empty boards to essentailly be in a valid finish state. An added complication of this function is now to ensure that all of the tiles on a board are used, by ensuring no tiles are unused. It is up to the responsibility of a game to ensure that a board has been given tiles before it checked to see if it is in a state to check words.