fixing shuffle
The tlies shuffle randomly, as expected.
Previously, they were shuffling using the same seed.  This meant that tiles were distributed in the same order at the start of each game.
Now, the shuffling is seeded with the time the server starts.  This is ok for the game, but the token key generator still uses a cryptographically-secure key generator.