not doing template TODO
Not checking for an error when rendering templates. The integration tests ensure the templates write correctly.
This could be done in a Server unit test, but it is somewhat pointless because it requires tring to render a template that WILL fail to execute and it is innenficient because to do so correctly, the template would have to be written to a buffer, then rewritten to the response.
simplifying code in getHandler()