using new for zero-valued references
Using new keyword when a zero value reference is needed.
Omitting assignment to empty struct when zero value is needed. Nil maps/slices have zero length, are ok for iteration. Using the make builtin when the maps should be empty, but will be populated later or will be compared to a non-nil (empty) value.

The only non-test files modified here are server/server.go and ui/game/canvas/canvas.go:
 * selver tls config fields set in struct creation
 * canvas start/end not set when empty pixel position value is desired