remove shared-mime-info dependency
Not checking some mime types in unit tests (.ico, LICENSE (plain)).  These mime types will also not be present on the computer that runs the server executable because /etc/mime.types is not bundled.
It is currently not critical to serve the license or favicon with the correct MIME type in the Content-Type header.  This package also adds about 11MB more to the build image (~2.5% larger).